+++ client/gutenberg/extensions/phone/edit.js
@@ -10,28 +10,9 @@ import { Component } from '@wordpress/element';
  * Internal dependencies
  */
 import { __ } from 'gutenberg/extensions/presets/jetpack/utils/i18n';
-import textMatchReplace from 'gutenberg/extensions/presets/jetpack/utils/text-match-replace';
+import save from './save';
 
-export function renderPhone( inputText ) {
-	return textMatchReplace(
-		inputText,
-		/([0-9\()+]{1}[\ \-().]?[0-9]{1,6}[\ \-().]?[0-9]{0,6}[\ \-()]?[0-9]{0,6}[\ \-().]?[0-9]{0,6}[\ \-().]?[0-9]{0,6}[\ \-().]?[0-9]{0,6})/g,
-		( number, i ) => {
-			if ( number.trim() === '' ) {
-				return number;
-			}
-			const just_number = number.replace( /\D/g, '' );
-
-			return (
-				<a itemprop="telephone" content={ just_number } href={ `tel:${ just_number }` } key={ i }>
-					{ number }
-				</a>
-			);
-		}
-	);
-}
-
-export class PhoneEdit extends Component {
+class PhoneEdit extends Component {
 	constructor( ...args ) {
 		super( ...args );
 		this.onChange = this.onChange.bind( this );
@@ -44,14 +25,12 @@ export class PhoneEdit extends Component {
 	render() {
 		const {
 			attributes: { phone },
-			className,
 			isSelected,
 		} = this.props;
 
 		return (
 			<div className={ isSelected ? 'jetpack-phone-block is-selected' : 'jetpack-phone-block' }>
-				{ ! isSelected &&
-					phone !== '' && <div className={ className }>{ renderPhone( phone ) }</div> }
+				{ ! isSelected && phone !== '' && save( this.props ) }
 				{ ( isSelected || phone === '' ) && (
 					<PlainText
 						value={ phone }
@@ -63,3 +42,5 @@ export class PhoneEdit extends Component {
 		);
 	}
 }
+
+export default PhoneEdit;
